Add feature for the leader recipe
