+++ Cargo.toml
@@ -21,17 +21,18 @@ edition = "2024"
 maintenance = { status = "experimental" }
 
 [dependencies]
-backon = "1.6.0"
+backon = { version = "1.6.0", optional = true }
 byteorder = "1.5.0"
 futures = "0.3.31"
 pin-project = "1.1.10"
 snafu = "0.8.6"
-tokio = { version = "1.47.1", features = ["net", "rt", "time", "sync"] }
+tokio = { version = "1.47.1", features = ["net", "rt", "time"] }
 tracing = "0.1.41"
-uuid = { version = "1.23.1", features = ["v4"] }
+uuid = { version = "1.23.1", features = ["v4"], optional = true }
 
 [dev-dependencies]
 tokio = { version = "1.47.1", features = ["macros"] }
 tracing-subscriber = "0.3.19"
 
 [features]
+leader_recipe = ["dep:backon", "dep:uuid", "tokio/sync"]
